Move TrackerNet enums
