+++ lib/enums/trackerNet/stations.ts
